+++ scripts/deploy.sh
@@ -5,18 +5,13 @@
 # that is needed to be flushed before deployment
 #
 
-if [ "$#" -ne 1 ]; then
-    echo "SQL version string must be provided as argument!"
-    exit 1
-fi
-
 #
 # Get current directory
 # @source https://stackoverflow.com/questions/59895/getting-the-source-directory-of-a-bash-script-from-within
 #
 function current_dir(){
     local SOURCE="${BASH_SOURCE[0]}"
-    while [ -h "$SOURCE" ]; do # resolve $SOURCE until the file is no longer a symlink
+    while [[ -h "$SOURCE" ]]; do # resolve $SOURCE until the file is no longer a symlink
       local DIR="$( cd -P "$( dirname "$SOURCE" )" && pwd )"
       SOURCE="$(readlink "$SOURCE")"
       [[ ${SOURCE} != /* ]] && SOURCE="$DIR/$SOURCE" # if $SOURCE was a relative symlink, we need to resolve it relative to the path where the symlink file was located
@@ -34,21 +29,33 @@ function current_dir(){
 #
 MYSQL_DB="adact"
 MYSQL_USER="root"
-MYSQL_PASS="root"
+MYSQL_PASS="PASSWORD"
 
 VERSION=$1
 DIR=$(current_dir)
 
 #
+# Run tests
+#
+python3 "$DIR/../tests/basic_test.py"
+if [[ $? -ne 0 ]]; then
+    exit 1
+fi
+python3 "$DIR/../tests/new_project_test.py"
+if [[ $? -ne 0 ]]; then
+    exit 1
+fi
+
+#
 # Remove files & directories
 #
-find "$DIR/../tmp/" \! -name 'readme.md' -delete
-find "$DIR/../Projects/" \! -name 'readme.md' -delete
+find "$DIR/../tmp/" \! -name 'readme.md' -delete 2> /dev/null
+find "$DIR/../Projects/" \! -name 'readme.md' -delete 2> /dev/null
 
 #
 # Flash mysql data
 #
-mysql --user=${MYSQL_USER} --password=${MYSQL_PASS} --database=${MYSQL_DB} << END
+mysql --user=${MYSQL_USER} --password=${MYSQL_PASS} --database=${MYSQL_DB} 2> /dev/null << END
 DELETE FROM active_sessions;
 DELETE FROM login_attempts;
 DELETE FROM pending_projects;
@@ -60,7 +67,9 @@ END
 #
 # Dump mysql database
 #
-mysqldump --user=${MYSQL_USER} --password=${MYSQL_PASS} ${MYSQL_DB} > "${DIR}/../sql/${MYSQL_DB}_v${VERSION}.sql"
+if [[ $# -eq 1 ]]; then
+    mysqldump --user=${MYSQL_USER} --password=${MYSQL_PASS} ${MYSQL_DB} > "${DIR}/../sql/${MYSQL_DB}_v${VERSION}.sql" 2> /dev/null
+fi
 
 #
 # Switch to the default configurations
